requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
python-dotenv==1.0.0
